Clean up matrix printing in qvm-app
I *think* the new stuff is equivalent to the old. The old stuff was
doing a lot of matrix calculations and rescaling and etc. There are
functions now that do all of this for us (gate-apps-to-log-mat,
scale-out-matrix-phases, etc)

I borrowed the matrix equality stuff from cl-quil-tests. Couple
questions,
  1. Is this all equivalent?
  2. Should matrix equality stuff live elsewhere (so that it's not
  duplicated across quilc/cl-quil/cl-quil-tests)
  3. Should that place be magicl?

Have a good weekend xxx